adding token validation to users page
git-svn-id: http://svn.concrete5.org/svn/concrete5@315 b0551a0c-1e16-4222-a7d5-975db1aca215

Former-commit-id: 721dff7cf287d6c1899d01501e0a32b65e693df0

+++ trunk/web/concrete/controllers/dashboard/users/controller.php
@@ -9,7 +9,7 @@ class DashboardUsersController extends Controller {
 	}
 	
 	
-	public function delete($delUserId){
+	public function delete($delUserId, $token = null){
 		$u=new User();
 		try {
 
@@ -35,6 +35,11 @@ class DashboardUsersController extends Controller {
 				throw new Exception(t('Invalid user ID.'));
 			}
 
+			$valt = Loader::helper('validation/token');
+			if (!$valt->validate('delete_account', $token)) {
+				throw new Exception($valt->getErrorMessage());
+			}
+			
 			$delUI->delete(); 
 			$resultMsg=t('User deleted successfully.');
